+++ example/lib/requests/request.dart
@@ -1,3 +1,0 @@
-import 'package:mvc_rocket/mvc_rocket.dart';
-
-final RocketRequest baseRequest = Rocket.get<RocketRequest>(rocketRequestKey);