+++ lib/src/mc_request.dart
@@ -106,7 +106,7 @@ class RocketRequest {
       if (model != null) {
         if (result is List?) {
           model.setMulti(result ?? []);
-          return model.multi;
+          return model.all;
         } else {
           model.fromJson(result);
           return model;
@@ -115,9 +115,11 @@ class RocketRequest {
         return result;
       }
     } else {
-      model!.state = RocketState.failed;
+      model!.setException(RocketException(
+        response: utf8.decode(await response.stream.toBytes()),
+        statusCode: response.statusCode,
+      ));
       _getDebugging(response, endpoint);
-      throw Exception('Failed to load Data');
     }
   }