+++ CHANGELOG.md
@@ -86,3 +86,4 @@
 - Removed `multi` parameter from RocketRequest method
 - Added end2end test on example
 - Added `rebuildFromSubModel` parameter on `RocketView` for control rebuild widget from sub-models
+- Renamed `multi` `RocketModel` field to `all`